requirements.txt
PyQt6
SpeechRecognition
pyttsx3
spotipy
pyautogui
psutil
python-dotenv
Pillow
pyaudio
requests
pywin32
